+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Change Log
 - Add hidden “Reveal in Finder” menu item to syntax style action menu in Preferences (visible with `Option` key).
 - Close Preferences window with esc key.
 - Improve find result panel behavior with Spaces.
+- Rename “Script Error Panel” to “Console Panel.”
 - Delay timing to save text key bindings setting.
 - Tweak window size sample window.